rewrite: custom iterator, support immediate break
Earlier, due to it being an async iterator itself, upon being "`break`-en" from outside (its consumer), it couldn't have broken (its input iterators) free immediately because it would still have been `await`-ing on a ".next()" promise of one of the iterators.

I.e.:

```
for await (const data of merged) {
   break // <- this wouldn't actually break until the *next* iteration
   // in some cases this could potentially mean indefinitely (e.g. process.stdin)
}
```

Now it creates the iterator itself so that it has more control over when can it break the iterators.